feat(auth0): move provider to its file
