+++ src/Auth0ProviderWithNavigate.tsx
@@ -1,0 +1,36 @@
+import { useNavigate } from 'react-router';
+import { Auth0Provider } from '@auth0/auth0-react';
+
+type AppState = {
+  returnTo?: string;
+  [key: string]: string | undefined;
+};
+
+export const Auth0ProviderWithNavigate = ({ children }: { children: React.ReactNode }) => {
+  const navigate = useNavigate();
+
+  const domain = import.meta.env.VITE_AUTH0_DOMAIN;
+  const clientId = import.meta.env.VITE_AUTH0_CLIENT_ID;
+  const redirectUri = import.meta.env.VITE_AUTH0_CALLBACK_URL;
+
+  const onRedirectCallback = (appState: AppState | undefined) => {
+    navigate(appState?.returnTo || window.location.pathname);
+  };
+
+  if (!(domain && clientId && redirectUri)) {
+    return null;
+  }
+
+  return (
+    <Auth0Provider
+      domain={domain}
+      clientId={clientId}
+      authorizationParams={{
+        redirect_uri: redirectUri,
+      }}
+      onRedirectCallback={onRedirectCallback}
+    >
+      {children}
+    </Auth0Provider>
+  );
+};
