+++ src/main.tsx
@@ -2,12 +2,12 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { BrowserRouter, useNavigate } from 'react-router';
+import { BrowserRouter } from 'react-router';
 import DugaRoutes from './routes/index.tsx';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { SocketProvider } from './context/SocketProvider.tsx';
-import { Auth0Provider } from '@auth0/auth0-react';
+import { Auth0ProviderWithNavigate } from './Auth0ProviderWithNavigate.tsx';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -19,42 +19,6 @@ const queryClient = new QueryClient({
     },
   },
 });
-
-type AppState = {
-  returnTo?: string;
-  [key: string]: string | undefined;
-};
-
-const Auth0ProviderWithNavigate = ({ children }: { children: React.ReactNode }) => {
-  const navigate = useNavigate();
-
-  const domain = import.meta.env.VITE_AUTH0_DOMAIN;
-  const clientId = import.meta.env.VITE_AUTH0_CLIENT_ID;
-  const redirectUri = import.meta.env.VITE_AUTH0_CALLBACK_URL;
-
-  console.log({ domain, clientId, redirectUri });
-
-  const onRedirectCallback = (appState: AppState | undefined) => {
-    navigate(appState?.returnTo || window.location.pathname);
-  };
-
-  if (!(domain && clientId && redirectUri)) {
-    return null;
-  }
-
-  return (
-    <Auth0Provider
-      domain={domain}
-      clientId={clientId}
-      authorizationParams={{
-        redirect_uri: redirectUri,
-      }}
-      onRedirectCallback={onRedirectCallback}
-    >
-      {children}
-    </Auth0Provider>
-  );
-};
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
